Edit on environment switching

+++ tests/PU.Users.Tests/WebApplicationFactory.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
@@ -10,37 +6,19 @@ using PU.Users.Api.Data;
 
 public class CustomWebApplicationFactory : WebApplicationFactory<Program>
 {
-    protected override void ConfigureWebHost(Microsoft.AspNetCore.Hosting.IWebHostBuilder builder)
+    protected override void ConfigureWebHost(IWebHostBuilder builder)
     {
         builder.ConfigureServices(services =>
         {
-            // Remove existing AppDbContext registration
-            var descriptor = services.SingleOrDefault(
-                d => d.ServiceType == typeof(DbContextOptions<AppDbContext>));
-            if (descriptor != null)
-                services.Remove(descriptor);
+            // Only reset DB if we're in Test environment
+            using var scope = services.BuildServiceProvider().CreateScope();
+            var env = scope.ServiceProvider.GetRequiredService<IWebHostEnvironment>();
 
-            // Add InMemory DB
-            services.AddDbContext<AppDbContext>(options =>
-            {
-                options.UseInMemoryDatabase("IntegrationTestDb");
-            });
-
-            // Build the service provider
-            var sp = services.BuildServiceProvider();
-
-            using (var scope = sp.CreateScope())
+            if (env.EnvironmentName == "Test")
             {
                 var db = scope.ServiceProvider.GetRequiredService<AppDbContext>();
                 db.Database.EnsureDeleted();
                 db.Database.EnsureCreated();
-
-                // Seed at least one group
-                if (!db.Groups.Any())
-                {
-                    db.Groups.Add(new PU.Users.Api.Models.Group { Name = "Test Group" });
-                    db.SaveChanges();
-                }
             }
         });
     }
